display the player inventory in the game interface
The "Cimoj" text was removed from the interface because it was pretty
pointless; it just took up space without adding values.

In the future, when a version number means something it may make sense
to center the program name and version on the top or bottom of the
screen. Something like the following
        cimoj v0.1.0 -- MIT License (c) 2016 Bryan Matsuo